[CARBONDATA-3636]Timeseries query is not hitting datamap if granularity in query is given case insensitive
Problem:
TimeSeriesUDF function has two parameters:

1.Column -> AttributeReference
2.Granularity -> Literal
For timeseries function having granularity with different case letters in Create datamap and actual query,
select query is not hitting the datamap. This is because, since we store granularity as Literal,
semanticEquals returns false, if granularity is of different case.

Solution:
For TimeSeriesFunction, check Literal value case insensitive

This closes #3541