[CARBONDATA-3631] StringIndexOutOfBoundsException When Inserting Select From a Parquet Table with Empty array/map
Modification reason:
(1) StringIndexOutOfBoundsException When Inserting Select From a Parquet Table with Empty array/map.
(2) ArrayIndexOutOfBuoundsException When Inserting Select From a Parquet Table with a map with empty key and empty value
(3) Result is incorrect when Inserting Select From a Parquet Table with a Struct with Empty String, The result will be null while the correct result is "".

Modification content:
(1) When the input value is ARRAY(), return EMPTY_DATA_RETURN in the FieldConverter.scala, ArrayParserImpl handle it.
(2) When the input value is ARRAY(""), return EMPTY STRING ->"" in the FieldConverter.scala, ArrayParserImpl handle it.
(3) When the input value is MAP("",""), return EMPTY STRING ->"" in the FieldConverter.scala, MapParserImpl handle it.
(4) When the input value is MAP(), return EMPTY_DATA_RETURN ->"" in the FieldConverter.scala, MapParserImpl handle it.

This closes #3545