[CARBONDATA-3600]Fix the cleanup failure issue if user fails to access table
Why is this PR needed?

the code to clean up the stale datamap folders during session initialization causes problem, if some other user tries to access the datamap table, we might get permission exception.

What changes were proposed in this PR?

We need to clean up only if the table does not exists in hive metastore, but the schema exists, we get exception incase the table exists but other user, so in such case, no need to go ahead, we can just catch exception and go ahead. Same changes are proposed here.

Does this PR introduce any user interface change?

No

Is any new testcase added?

No (Not required as tested in cluster and fix is for clean up issue)

This closes #3548