[CARBONDATA-3644Support Configuration of Complex Delimiters in Carbon Properties
Modification reason: In the insert carbontable select from a parquet table processing, if the binary column has the content '\001', like 'col1\001col2', the content before '\001' will be truncated as '\001' is the Complex Delimiter. The problem is that Complex Delimiter can't be configured in the insert flow, which needs to improve.

Modification content: We add the configuration of complex delimiters in the CarbonProperties, which will be loaded in the LoadOption.