[CARBONDATA-3642] Add column name in error msg when string length exceed 32000
Problem: Currently when string length exceeds 32k for a column, it fails with exception, but it doesn't
give the column name and proper error.

Solution: When length exceeds 32k for column, throw exception with proper error message which contains
column name and suggestion to use long_string_columns table property.

This closes #3546