[CARBONDATA-3565] Fix complex binary data broken issue when loading dataframe data
Why is this PR needed?
When binary data is DataOutputStream#writeDouble and so on.
Spark DataFrame(SQL) load it to a table, the data will be broken (EF BF BD) when reading out.

What changes were proposed in this PR?
If data is byte[], no need to convert to string and decode to byte[]
again

Does this PR introduce any user interface change?
No

Is any new testcase added?
Yes

This closes #3430