[HOTFIX] Fix INSERT STAGE footer read error
Why is this PR needed?
There are 2 issues:
1. INSERT STAGE will skip bytes to read the last long in the carbondata file as
the offset to the footer, but sometimes DFS is not skipping the right byte size
2. When error occurs, segment status is not in SUCCESS state, but recovery
handling is not doing correctly in INSERT STAGE command

What changes were proposed in this PR?
1. In CarbonFooterReaderV3.readFooterVersion3, if offset is not set by
caller, read it from corresponding index file.
2. In CarbonInsertFromStageCommand, when recovery is required, delete the
segment entry if the entry is not in SUCESS state.

This closes #3561