[CARBONDATA-3663] Support loading stage files in batches
Why is this PR needed?
When there are a lots of stage files in the stage directory, if load all of them in once time, the loading time will can not be control.
There need a way for users to specify the number of stage files per processing, to control the execution time of commands.

What changes were proposed in this PR?
Add a load option batch_file_count for users to specify the number of stage files per processing.

Does this PR introduce any user interface change?
Yes

Is any new testcase added?
Yes

This closes #3578