[CARBONDATA-3666] Avoided listing of table dir in refresh command
Why is this PR needed?
Currently if a refresh command is fired on a parquet table using carbon session then carbon will list all the tables and check whether the table exists or not, then we check if the schema file exists or not by listing the Metadata folder. This can be a problem in cloud scenarios as the listing on S3 is slow.

What changes were proposed in this PR?
get the metadata for the specified table, Then go for table listing only if the provider is carbon or the table is not registered in hive

Does this PR introduce any user interface change?
No

Is any new testcase added?
No

This closes #3581