[CARBONDATA-3676] Support clean carbon data files of stages.
Why is this PR needed?

At the end of the CarbonInsertFromStageCommand, the stage files will be cleared, but the data files which referenced by stage files will be not cleared. This could lead to a large backlog of data files。

What changes were proposed in this PR?

Provide a new command to allows us to delete data files which referenced by disabled table stages.
The new command is CarbonDeleteStageCommand.

Does this PR introduce any user interface change?
Yes

Is any new testcase added?
Yes

This closes #3602