[CARBONDATA-3680] core changes and load changes for SI integration
Why is this PR needed?
Currently we have datamaps like,* default datamaps* which are block and
blocklet and coarse grained datamaps like bloom, and fine grained
datamaps like lucene which helps in better pruning during query. What if we
introduce another kind of datamap which can hold blockletId as index? Initial level,
we call it as index which will work as a child table to the main table like we have
MV in our current code.

Yes, lets introduce the secondary index to carbon table which will be the
child table to main table and it can be created on column like we create
lucene datamap, where we give index columns to create index. In a similar way,
we create secondary index on column, so indexes on these column will be blocklet IDs
which will help in better pruning and faster query when we have a filter query on the
index column.

What changes were proposed in this PR?
introduced SI feature
it contains:

create SI table
load to SI
query from SI

This closes #3608