[CARBONDATA-3696] Avoid list db's all Tables to check table exists in the db
Why is this PR needed?

I found a lot of listTables of the db to check if table exists like this:

sparkSession.sessionState.catalog.listTables(databaseName)
  .exists(_.table.equalsIgnoreCase(tableName)
This may affect performances if there are lots of tables in one database. In spark, we can check this by tableExists function.

What changes were proposed in this PR?
use exits(db, table) to replace iterate tables in a db.

Does this PR introduce any user interface change?
No

Is any new testcase added?
No

This closes #3631