[CARBONDATA-3709] Move stage_data directory to the $tablePath/ from $tablePath/Metadata/
Why is this PR needed?
The stage_data directory is data directory, but under the $tablePath/Metadata/ directory, it cause the metadata directory is not clean.

What changes were proposed in this PR?
Move stage_data directory to the $tablePath/ from $tablePath/Metadata/, and keep the $tablePath/Metadata/ directory not include any data.

This closes #3626