[CARBONDATA-3723] Fix insert failure when partition column is a first column of projection in MV
Why is this PR needed?
For Alter table drop and add column scenarios, partition columns
may not be in the end of create-order column. So, added a logic
to keep the partition column in the end in #3634.

But partition may not be the last projection column during mv creation.
So now, create order partition is kept last, but the projection is not last.
Hence the insert gives wrong data or fails.

What changes were proposed in this PR?
rearrage partition to last in create order columns, only for alter table scenarios

This closes #3640