[CARBONDATA-3716] Fixed spark 2.4 UT failures
Why is this PR needed?

Issue:
a) In 2.4 subquery alias produces an output set with alias identifier(that is both database_name and table_name) which causes failures.
Solution- Get subquery alias output with only table_name as alias
b) In 2.4 a new variable was introduced "spark.sql.legacy.allowCreatingManagedTableUsingNonemptyLocation" which is by default set to "false".
While refreshing table In 2.4 we need to set this variable to "true" in order to create table in non empty location.

Does this PR introduce any user interface change?
No

Is any new testcase added?
No

This closes #3629