[CARBONDATA-3732] Fix ClassCastException with rand() udf function
Why is this PR needed?
ClassCastException is thrown, while running queries with cast of rand() function. This is because, while making a deterministic expression, we are converting an alias of CustomDeterministic expression to CustomDeterministic expression again, which is not needed.

What changes were proposed in this PR?
Do not convert an Alias of CustomDeterministic expression to CustomDeterministic expression again

Does this PR introduce any user interface change?
No

Is any new testcase added?
Yes

This closes #3648