[CARBONDATA-3734] Fix insert failure on partition table when parition column is in sort column
Why is this PR needed?

When partition column is a sort column.

a) currently sort columns won't be the head of attributes, so need to add logic as per that, as partition columns will be in the end for global sort.

b) While rearranging the data fields at executor, need to keep partition column in the end even though it is in sort column.

What changes were proposed in this PR?
prepare sort columns for global sort based on the table sort columns.
Keep partition in the end, even though it is in sort columns for data field attributes

Does this PR introduce any user interface change?
No

Is any new testcase added?
No (already this scenario present in mv test case)

This closes #3654