[HOTFIX] Remove Useless filepath truncation During prunning
Why is this PR needed?
During prunning, it will truncate the filepath with the keyword "/Part0". But it is useless.
Specially, for partition table, the filepath is like '/tablename/partitionkey=value/segment0/.carbondata', it even have no keywork '/Part0'.
Then for non-partition table, it is also meaningless to do this work.
If there are a huge amount of blocklets, such as 400 million blocklets are going to do prunned, the truncation hampers the query performance heavily.

What changes were proposed in this PR?
Remove the filepath truncation.

Does this PR introduce any user interface change?
No

Is any new testcase added?
No

This closes #3660