[CARBONDATA-3751] Segments are not Marked for delete if everything is deleted in a segment with index server enabled
Why is this PR needed?
When all the rows are deleted from a segment with index server enabled, the segment is not Marked for delete.

What changes were proposed in this PR?
Rowcount was always being sent as 0 from the index server DistributedPruneRDD. Getting that value from DataMapRow now

Does this PR introduce any user interface change?
No

Is any new testcase added?
No

This closes #3679