[HOTFIX] Fix all flink test case failure and enable UT in CI
Why is this PR needed?
After #3628, default BATCH_FILE_ORDER is wrong [it is not ASC or DSC].
so, all the test case in flink module failed due as no order is set in stage command.
Also flink UT is not running in CI, hence it is not caught

What changes were proposed in this PR?
Fix the default value to Ascending order.
Enable UT running for flink module.

Does this PR introduce any user interface change?
No

Is any new testcase added?
No

This closes #3685