[CARBONDATA-3759] Refactor segmentRefreshInfo and fix cache issue in multiple application scenario
Why is this PR needed?
currently the segmentRefreshInfo is helping to clear the cache only in update cases and it fails to refresh the cache if any segment files changes or get updates.
when two applications are running on same store. One application changes some segment files changes and removes old cache and may be delete files, which wont be reflected in other application, which may result in either wrong results or query failure.

What changes were proposed in this PR?
refactor the segmentRefreshInfo to clear the cache and update when there are any updates on segments and if the segment files of respective segments changes.

Does this PR introduce any user interface change?
No

Is any new testcase added?
No
Tested in cluster and existing test cases will be enough.

This closes #3686