[CARBONDATA-3787] Fixed MemoryLeak in data load and compaction
Why is this PR needed?
PR #3638 uses direct byte buffer. This PR is for fixing Offheap memory leak during load/compaction

What changes were proposed in this PR?
DirectByteBuffer are cleaned in offheap memory by JVM, when GC happens (when heap is full). Consider a scenario, where heap is not full, but offheap is getting full. In this case DirectByteBuffers is not cleaned as GC not happened. This time we may get OOM from offheap memory and yarn may kill the jvm process. So, it is better to clean up the DirectByteBuffers after the usage by calling a reflection method.

Does this PR introduce any user interface change?
No

Is any new testcase added?
No

This closes #3706