[HOTFIX] Fixed task distribution issue in SegmentPruneRDD
Why is this PR needed?
SI queries are degraded because getPrefferedLocation is
not overridden in SegmentPruneRDD due to which tasks are fired randomly to any executors

What changes were proposed in this PR?
override getPrefferedLocation so that tasks are fired to correct executors.

Does this PR introduce any user interface change?
No

Is any new testcase added?
No

This closes #3712