[CARBONDATA-3773] Skip Validate partition info in Indexserver count star flow
Why is this PR needed?
BlockIndex.ValidatePartitionInfo check was added for IUD scenario as part of countStar query
performance improvement in PR-3148. For count(*) flow with indexserver, validatePartitionInfo
is called multiple times (No. of Blocks * No. of Partitions), which is not required.

What changes were proposed in this PR?
Skip ValidatePartition check in case of Count(*) query with indexserver

This coses #3714