[CARBONDATA-3788] Fix insert failure during global sort with huge data in new insert flow
Why is this PR needed?
Spark is resuing the internalRow in global sort partition flow with huge data.
As RDD of Internal row is persisted for global sort.

What changes were proposed in this PR?
Need to have a copy and work on the internalRow before the last transform for global sort partition flow.
Already this was doing for insert stage command (which uses global sort partition)

This closes  #3732