[CARBONDATA-3789] Fix cache issue in case of compaction failure in compaction post listeners
Why is this PR needed?
Consider a scenario where the post-event Is called in case of compaction and then the listener is basically selecting the ongoing compacted segment data and loading SI table load after main table compaction. During that time the cache is loaded, if in case this is failed, then the cache is still present but the actual segment isn't, so which will lead to consecutive failures.

What changes were proposed in this PR?
So, when the failure happens, the cache should be cleared either in the index server or in the driver cache.

Does this PR introduce any user interface change?
No

Is any new test case added?
No

This closes #3733