[CARBONDATA-3792]Refactor system folder location and removed unwanted property
Why is this PR needed?
Currently the carbon.system.folder.location is being used only in test code,
which is not required and since the system folder location is based on database
and that code is being duplicated.

What changes were proposed in this PR?
Refactor code to get the system folder location based on db
location without duplicate code and remove the carbon.system.folder.location
usage in all the test classes and remove unnecessary code.

This closes #3743