[CARBONDATA-3793]Data load with partition columns fail with InvalidLoadOptionException when load option 'header' is set to 'true'
Why is this PR needed?
Data load with partition columns fail with InvalidLoadOptionException when load option header
is set to true.

What changes were proposed in this PR?
In SparkCarbonTableFormat.prepareWrite() method after adding the fileheader option with header
columns to optionsFinal, need to set header option to false.

This closes #3745