[CARBONDATA-3817]Fix table creation with all columns as partition columns
Why is this PR needed?
When all the columns are given as partition columns during create table,
create table should fail as a minimum one column should be present as a
non-partition column. This is because after #3574 , we improved the create
data source table and we call CreateDataSourceTableCommand of spark.
Since we are creating as Hive table, if while creating hive compatible way,
if it fails, then it will fall back to save its metadata in the Spark SQL
specific way, so partition validation fails when we try to store in hive
compatible way, so in retry, it will pass which is wrong behavior for hive compatible table.
in Hive integration location do not have file system URI prepared for it

What changes were proposed in this PR?
For partition table, if all the columns are present as partition columns,
then validate with the same API which spark does. append file system URI
for location parameter while inferring schema.

This closes #3762