[CARBONDATA-3820] Fix CDC failure when sort columns present in source dataframe
Why is this PR needed?
While merging into table with sortcolumns in the CDC Flow. The following exception will be throwed:
"column: id specified in sort columns does not exist in schema".
Root cause is that we use TBLProperteis with sortcolumns to create the TUPLEID_statusOnMerge carbonwriter, in which the sortcolumn、sortscope are all useless.

What changes were proposed in this PR?
remove the sortcolumn property when creating TUPLEID_statusOnMerge carbonwriter.

Does this PR introduce any user interface change?
No

Is any new testcase added?
Yes

This closes #3764