[CARBONDATA-3833] Make geoID visible
Why is this PR needed?
To make geohash column visible to the user

What changes were proposed in this PR?
Generated geohash column is included in the schema. Validation added to avoid alter and drop the geohash column. Indexes, MV and other table properties not supported on this column.

Does this PR introduce any user interface change?
No

Is any new testcase added?
Yes

This closes #3774