[CARBONDATA-3948] Fix for Bloom Index create failure
Why is this PR needed?
In the case when the whole flow goes to fallback mode in the
index server, the parent table retrieved again in the processData
function of carboncreateIndexCommand does not contain the serialized
IndexMetaData. This issue is only seen in spark-sql because the same
jvm is used between driver and executor

What changes were proposed in this PR?
Setting the index information in parent table after the table is
retrieved again and creating it if it is null

This closes #3775