[CARBONDATA-3860] Fix IndexServer keeps loading some segments index repeatly
Why is this PR needed?
In current getTableBlockIndexUniqueIdentifiers function. if the segmentBlockIndexInfo.getSegmentMetaDataInfo() is null, the IndexServer will keeps loading the index of this segment repeatly. We shall avoid to let it affect query performance, considering MetaDataInfo doesn't matter with the query processing.

What changes were proposed in this PR?
Return the cached index directly if segmentBlockIndexInfo.getTableBlockIndexUniqueIdentifiers() is not empty

Does this PR introduce any user interface change?
NO

Is any new testcase added?
NO

This closes #3795