[CARBONDATA-3871] Optimize performance when getting row from heap
Why is this PR needed?
Currently carbon uses priority queue to sort holders of sorted rows.
It first polls a holder from the heap, and adds it back if holder is not empty.
This will cause two times heap maintainance. We can reduce half of that operation,
and test shows it can save one third of time to get rows .

What changes were proposed in this PR?
What will be done when poll item from priority queue currently is:
1. remove first item.
2. move the last item to the position of first item, siftDown the new first item.
In this patch, we will peek(without removing from heap) the first item and get a row,
and siftDown the holder to a proper position if the holder is not empty.
Since this will affect order of non-sorted columns, it depends on #3813

Does this PR introduce any user interface change?
No

Is any new testcase added?
No

This closes #3804