[CARBONDATA-3873] Secondary index compaction with maintable clean files causing exception
Why is this PR needed?
Compaction with secondary index and cleaning files of the main table after compaction fails then we are
getting Exception.
If any compaction is failed for one of SI and if it is not disable then it will be considered for query
execution. When query executed on SI table which is failed in compaction and it doesn't have correct data
hence received Exception.

What changes were proposed in this PR?
1. If any compaction is failed then disable all SI which are successful. They will be enabled after the next load.
2. Changing Error to the warning in table creation flow as it is a success case we should not give ERROR.

This closes #3808