[CARBONDATA-3900][CARBONDATA-3882][CARBONDATA-3881] Fix multiple concurrent issues in table status lock and segment lock for SI and maintable
Why is this PR needed?
[CARBONDATA-3900] Fix maintable load failure in concurrent load and compaction scenario
Main table load flow segment lock is released before updating the table status success.
So, concurrent operation was considering this segment as stale segment (as segment lock is not present) and cleaning it. Leading to unable to get file status exception.
[CARBONDATA-3882] Fix wrong lock and missing Table status lock in some SI flows
In updateLoadMetadataWithMergeStatus, we want update SI table status, but lock is acquired on main table
triggerCompaction, updateTableStatusForIndexTables -> table status write is happening without lock
[CARBONDATA-3881] Fix concurrent main table compaction and SI load issue
Consider a scenario, where segmentX has loaded to main table but failed to load to SI table. So, while loading another segmentY, we reload failed SI segmentX. this time if the segmentX is compacted in main table and clean files executed on it. SI load will fail and segment id will not be found in segmentMap of SI and it throws exception.

What changes were proposed in this PR?
for [CARBONDATA-3900]
release segment lock after main table is updated.
for [CARBONDATA-3882]
In updateLoadMetadataWithMergeStatus, take a lock on SI table
triggerCompaction, updateTableStatusForIndexTables -> add a lock for table status write
for [CARBONDATA-3881]
just before reloading the failed SI segment. check if it is valid segment in main table.

Does this PR introduce any user interface change?
No

Is any new testcase added?
No

This closes #3810