[CARBONDATA-3895] Fix FileNotFound exception in query after global sort compaction
Why is this PR needed?
After global sort compaction, if we execute clean files and run the query or update delete operations, we get file not found exceptions and some data lost. This is because we form new load model for global sort compaction and facttimestamp is not set, so carbondata files are generated with a timestamp as 0.

What changes were proposed in this PR?
copy the facttimestamp from the incoming loadmodel and set it into new load model

Does this PR introduce any user interface change?
No

Is any new testcase added?
Yes

This closes #3812