[Carbondata-3883] Added filtering for the deleted rows for local dictionary fields
Why is this PR needed?
In Presto integration when reading records for table with string fields and local dictionary enabled gives improper result.

What changes were proposed in this PR?
Added a CarbondataColumnVectorWrapper for the dictionary vector for column for which internally create dictionary vector.
Wrapper is used to filter the deleted records.

Does this PR introduce any user interface change?
No

Is any new testcase added?
No

This closes #3818