[CARBONDATA-3853] Data load failure when loading with bucket column as DATE data type
Why is this PR needed?
While loading data into a table which has bucket column as DATE data type then receiving data load failure due to class-cast exception. The class-cast exception is due to DATE data type is not going as IntegralHash type for getting the hash because of there is no case for DATE datatype and it is going for StringHash type and typecast results class-cast Exception.

What changes were proposed in this PR?
Added a OR (||) condition if data type is DATE then it will go IntegralHash getHash() method for getting the hash code.

Does this PR introduce any user interface change?
No

Is any new testcase added?
Yes

This closes #3830