[CARBONDATA-3852] [CARBONDATA-3851] Fix Data update issue with SCD/CDC for partition table On spark cluster deploy mode
Why is this PR needed?
Issue 1:
On process IUD with mergeDataSetCommand, we are setting direct write to store path carbon property
as true in executor side and resetting the value to false on driver side. Because of this, Update
operation is unsuccessfull, as Issue 2 will come into picture.

Issue 2:
Currently, For insert/load data into partition table, data file will be moved to partition path
and index files will remain in tmp file to improve commit job performance.
When direct write to store path is enabled, data files are not moved to partition path and load
metrics are not set and added to task context. On commitJob, it just return, as index and carbon
size will be 0. Thus, the corresponding load is not successful.

What changes were proposed in this PR?
For Issue 1:
Set direct write to store path carbon property to default value on close writer while processIUD
for merge data set command

For Issue 2:
1. While writing data file, check if partition table and dir path is "partition tmp path", then
write data file directly to partiiton path.
2. On close writer, add partition carbon data and index info to metrics

This closes #3835