[CARBONDATA-3927]Remove unwanted fields from tupleID to make it short and to improve store size and performance
Why is this PR needed?
Currently, we store the tupleId which is very long, which
increases the store size and reduces query performance.

What changes were proposed in this PR?
Remove the compressor name, part id and batch number from the tupleID.
This helps to improve the store size and improve the query performance also.

part_0 is not required as it will be same for all
_batchno is not required as common word, can be replaced by -
compressor name not required as it does not plan any role.

This closes #3837